refactor(models): fix every models's type imports
