+++ models/global-setting/types.d.ts
@@ -1,7 +1,6 @@
 import type { Document, Types } from 'mongoose';
 import type { IBaseModel } from '../types';
-import type { ISMTPMailerDoc } from '@models/smtp-mailer/types';
-import type { ISMTPProviderDoc } from '@models/smtp-provider/types';
+import type { ISMTPProviderDoc, ISMTPMailerDoc } from '@models/types';
 import type { ID } from '@typs/model.objectid';
 
 type OTHER = boolean | string | number;
