+++ models/otp/types.d.ts
@@ -1,6 +1,6 @@
 import type { Document, Types } from 'mongoose';
 import type { IBaseModel } from '../types';
-import type { IUserDoc } from '@models/user/types';
+import type { IUserDoc } from '@models/types';
 import type { ID } from '@typs/model.objectid';
 
 export interface IOtp {
