+++ models/scope/types.d.ts
@@ -1,7 +1,6 @@
 import type { Document, Types } from 'mongoose';
 import type { IBaseModel } from '../types';
-import type { IFrontendDoc } from '@models/frontend/types';
-import type { ICredentialsDoc } from '@models/credential/types';
+import type { ICredentialsDoc, IFrontendDoc } from '@models/types';
 import type { ID } from '@typs/model.objectid';
 
 export interface IScope {
