+++ models/smtp-mailer/types.d.ts
@@ -1,6 +1,6 @@
 import type { Document, Types } from 'mongoose';
 import type { IBaseModel } from '../types';
-import type { ISMTPProviderDoc } from '@models/smtp-provider/types';
+import type { ISMTPProviderDoc } from '@models/types';
 import type { ID } from '@typs/model.objectid';
 
 export interface ISMTPMailer {
