+++ models/types.d.ts
@@ -9,3 +9,20 @@ export interface IBaseModel<U extends Document> extends Model<U> {
    */
   clearAll: (this: IBaseModel) => Promise<IInlineResponse<string>>;
 }
+
+export * from './blacklisted-user/types';
+export * from './credential/types';
+export * from './frontend/types';
+export * from './global-setting/types';
+export * from './otp/types';
+export * from './pending-user/types';
+export * from './policy/types';
+export * from './role/types';
+export * from './scope/types';
+export * from './service-account/types';
+export * from './session/types';
+export * from './smtp-mailer/types';
+export * from './smtp-provider/types';
+export * from './templates/types';
+export * from './tokens/types';
+export * from './user/types';
