+++ models/user/types.d.ts
@@ -1,8 +1,6 @@
 import type { Document, Types } from 'mongoose';
 import type { IBaseModel } from '../types';
-import type { IRoleDoc } from '@models/role/types';
-import type { IPolicyDoc } from '@models/policy/types';
-import type { IScopeDoc } from '@models/scope/types';
+import type { IScopeDoc, IPolicyDoc, IRoleDoc } from '@models/types';
 import type { ID } from '@typs/model.objectid';
 
 export interface IUser {
